Fix CI failure
First, it's weird that the latest revision passed CI and then started
failing on scheduled builds and pull requests.

Second, I don't want to waste time adopting or configuring somebody
else's cosmetic prefernce such as:

53 |     return Err(format!("unknown key '{}'", input).into());
   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: remove `return`: `Err(format!("unknown key '{}'", input).into())`

which is causing the current CI failure. It's not really detecting a
"mistake", which cargo clippy claims to point out.