Allow matching any key
Current configuration is allowing to match a key and remap it. While it is useful
for must cases, it is less useful when trying to disable shortcuts.
If there are a lot of shortcuts to disable, it may be easier (and better
for security) to disable all shortcuts except some.

This PR is adding a new "virtual" key used to match all keys. For instance, using:
```
keymap:
- remap:
    a: b
    ANY: null
```
will result in disabling all keyboards keys except ``KEY_A``, which is remapped to
``KEY_B``.

In case of shortcuts, this will disable all ``C-$KEY`` except ``C-p``:
```
keymap:
- exact_match: true
  remap:
    C-p: C-p
    C-ANY: reserved
```

Signed-off-by: Arnaud Patard <arnaud.patard@collabora.com>