Allowing matching event device on udev properties
This is a follow-up of PR584. When the device IDs are not known, it's still
possible to match them if they are sharing a common udev property (as shown with
``udevadm info /dev/input/eventX``).

So, add the udev crate and use it to get the device property to match.
Given that it may not be desirable for everyone, this has been put
inside a feature. Moreover, this will avoid build issues in the CI,
due to the musl toolchain missing the udev development files.

Signed-off-by: Arnaud Patard <arnaud.patard@collabora.com>