Allow any action in modmap's alone parameter
I wanted to configure the meta key to open an application launcher if
pressed on its own, but act normally if used as part of another
keybinding. I tried to use the held/alone options for that, but the
launch action isn't supported there.

This PR extends the alone option to match how the press option works.
That means any KeymapAction can be used instead of only the
ModmapActions.

I think this is technically a breaking change because existing configs
will have a slightly different keypress order when mapping one key to
many in the alone option. For example, `alone: [a, b]` used to trigger
keydown for a and b, then keyup for those two. Now it'll press and
release a, then press and release b. The original behavior can be
emulated with something like the following (assuming there aren't
subtleties I'm missing):
`alone: [{press: a}, {press: b}, {release: b}, {release: a}]`. I'm not
sure if this is something people are relying on or not though.