Fix bug with held keys
The previous commit introduced a bug because it tried to use the action
dispatcher to release the held keys instead of using the return value. I
thought this would be equivalent, but this introduces a timing problem
where the key is released before it's pressed. That causes the press to
become sticky.

I'm solving it in this commit by returning the held keys to their
original code path. Those will be handled exactly as before. The alone
actions are now the only ones using the dispatch method.